backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0
easyocr==1.7.1
langdetect==1.0.9
firebase-admin==6.2.0
numpy==1.24.3
python-dotenv==1.0.0
sse-starlette==1.6.5